src/requirements.txt
annotated-types==0.5.0
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
paho-mqtt==1.6.1
pydantic==2.0.3
pydantic_core==2.3.0
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.4
